corr-index

+++ src/index.ts
@@ -24,10 +24,12 @@ app.use((err: Error, req: Request, res: Response, next: NextFunction): express.R
 })
 app.use('/api', routes)
 
-db.sync({ force: false })
-  .then(() => {
-    app.listen(3001, () => console.log('Servidor en el puerto 3001'))
-  })
-  .catch(console.error)
+if (require.main === module) {
+  db.sync({ force: false })
+    .then(() => {
+      app.listen(3001, () => console.log('Servidor en el puerto 3001'));
+    })
+    .catch(console.error);
+}
 
 export default app
